requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-import-export==2.3.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
tablib==2.0.0
ujson==3.1.0
urllib3==1.25.10
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
